+++ NEWS.md
@@ -1,4 +1,9 @@
 # cvpiaHabitat 1.0.0
-### Initial non-development release 6/12/2018
+Initial non-development release 6/12/2018
 # cvpiaHabitat 1.0.1
-### Adjust Mokelumne Spawning Data 10/30/2018
+Adjust Mokelumne Spawning Data 10/30/2018
+# cvpiaHabitat 1.0.2 (in progress)
+Adjust Stanislaus rearing habitat for steelhead
+Add vignette describing floodplain habitat scaling
+Mokelumne
+American spawning WUA updated